Moved ElemenList component to Formik folder
