+++ src/containers/ConceptPage/components/ConceptArticles.tsx
@@ -9,7 +9,7 @@ import React, { FC, useState } from 'react';
 import { injectT, tType } from '@ndla/i18n';
 import { FieldHeader } from '@ndla/forms';
 import { FormikHelpers, FormikValues } from 'formik';
-import ElementList from '../../NdlaFilm/components/ElementList';
+import ElementList from '../../FormikForm/components/ElementList';
 import { AsyncDropdown } from '../../../components/Dropdown';
 import {
   ArticleType,
