+++ src/containers/FormikForm/components/ElementListItem.tsx
@@ -13,8 +13,8 @@ import styled from '@emotion/styled';
 import { spacing, colors, fonts, animations } from '@ndla/core';
 import Tooltip from '@ndla/tooltip';
 import { DragHorizontal, DeleteForever } from '@ndla/icons/editor';
-import { resourceToLinkProps } from '../../../util/resourceHelpers';
-import { ContentResultType } from '../../../interfaces';
+import { resourceToLinkProps } from '../../util/resourceHelpers';
+import { ContentResultType } from '../../interfaces';
 
 const ELEMENT_HEIGHT = 69;
 const ELEMENT_MARGIN = 4;
